+++ odbc-api/src/environment.rs
@@ -1,6 +1,8 @@
+use std::{cmp::max, collections::HashMap};
+
 use crate::{handles, Connection, Error};
-use odbc_sys::AttrOdbcVersion;
-use widestring::{U16Str, U16String};
+use odbc_sys::{AttrOdbcVersion, FetchOrientation};
+use widestring::{U16CStr, U16Str, U16String};
 
 /// An ODBC 3.8 environment.
 ///
@@ -153,5 +155,113 @@ impl Environment {
         let mut connection = self.environment.allocate_connection()?;
         connection.connect_with_connection_string(connection_string)?;
         Ok(Connection::new(connection))
+    }
+
+    /// Get information about available drivers. Only 32 or 64 Bit drivers will be listed, depending
+    /// on wether you are building a 32 Bit or 64 Bit application.
+    ///
+    /// # Example
+    ///
+    /// ```no_run
+    /// use odbc_api::Environment;
+    ///
+    /// let mut env = unsafe { Environment::new () }?;
+    /// for driver_info in env.drivers()? {
+    ///     println!("{:#?}", driver_info); 
+    /// }
+    ///
+    /// # Ok::<Result<_, odbc_api::Error>>(())
+    /// ```
+    pub fn drivers(&mut self) -> Result<Vec<DriverInfo>, Error> {
+        // Find required buffer size to avoid truncation.
+
+        // Start with first so we are independent of state
+        let (mut desc_len, mut attr_len) = if let Some(v) = self
+            .environment
+            .drivers_buffer_len(FetchOrientation::First)?
+        {
+            v
+        } else {
+            // No drivers present
+            return Ok(Vec::new());
+        };
+
+        // If there are let's loop over the rest
+        while let Some((cand_desc_len, cand_attr_len)) = self
+            .environment
+            .drivers_buffer_len(FetchOrientation::Next)?
+        {
+            desc_len = max(cand_desc_len, desc_len);
+            attr_len = max(cand_attr_len, attr_len);
+        }
+
+        // Allocate +1 character extra for terminating zero
+        let mut desc_buf = vec![0; desc_len as usize + 1];
+        let mut attr_buf = vec![0; attr_len as usize + 1];
+
+        let mut driver_infos = Vec::new();
+        while self.environment.drivers_buffer_fill(
+            FetchOrientation::Next,
+            &mut desc_buf,
+            &mut attr_buf,
+        )? {
+            let description = U16CStr::from_slice_with_nul(&desc_buf).unwrap();
+            let attributes = U16CStr::from_slice_with_nul(&attr_buf).unwrap();
+
+            let description = description.to_string().unwrap();
+            let attributes = attributes.to_string().unwrap();
+            let attributes = attributes_iter(&attributes).collect();
+
+            driver_infos.push(DriverInfo{ 
+                description,
+                attributes,
+            });
+        }
+
+        Ok(driver_infos)
+    }
+}
+
+/// Struct holding information available on a driver. Can be obtained via [`Environment::drivers`].
+#[derive(Clone, Debug, Eq, PartialEq)]
+pub struct DriverInfo {
+    /// Name of the ODBC driver
+    pub description: String,
+    /// Attributes values of the driver by key
+    pub attributes: HashMap<String, String>,
+}
+
+/// Called by drivers to pares list of attributes
+///
+/// Key value pairs are separated by `\0`. Key and value are separated by `=`
+fn attributes_iter(attributes: &str) -> impl Iterator<Item = (String, String)> + '_ {
+    attributes
+        .split('\0')
+        .take_while(|kv_str| *kv_str != String::new())
+        .map(|kv_str| {
+            let mut iter = kv_str.split('=');
+            let key = iter.next().unwrap();
+            let value = iter.next().unwrap();
+            (key.to_string(), value.to_string())
+        })
+}
+
+#[cfg(test)]
+mod test {
+
+    use super::*;
+
+    #[test]
+    fn parse_attributes() {
+        let buffer = "APILevel=2\0ConnectFunctions=YYY\0CPTimeout=60\0DriverODBCVer=03.\
+                      50\0FileUsage=0\0SQLLevel=1\0UsageCount=1\0\0";
+        let attributes : HashMap<_,_>= attributes_iter(buffer).collect();
+        assert_eq!(attributes["APILevel"], "2");
+        assert_eq!(attributes["ConnectFunctions"], "YYY");
+        assert_eq!(attributes["CPTimeout"], "60");
+        assert_eq!(attributes["DriverODBCVer"], "03.50");
+        assert_eq!(attributes["FileUsage"], "0");
+        assert_eq!(attributes["SQLLevel"], "1");
+        assert_eq!(attributes["UsageCount"], "1");
     }
 }